Bugfixing ingest process

+++ db/migrate/20140118010609_add_progress_to_ingests.rb
@@ -1,5 +1,5 @@
 class AddProgressToIngests < ActiveRecord::Migration
   def change
-    add_column :ingests, :progress, :integer, :default => 0, :null => false
+    add_column :ingests, :progress, :float, :default => 0.0, :null => false
   end
 end
